backend/requirements.txt
fastapi
uvicorn
pandas
scikit-learn
python-multipart
google-genai
pydantic
